Fix for sitelink $0 price updates (manager special) Support for Google maps premier API

+++ grails-app/controllers/storitz/StorageSiteController.groovy
@@ -471,7 +471,7 @@ class StorageSiteController {
     if (params.rentalTransactionId) {
       rentalTransactionInstance = RentalTransaction.get(params.rentalTransactionId as Long)
     }
-    StorageSite site = StorageSite.get(params.id)
+    StorageSite site = StorageSite.get(params.id as Long)
 
     if (site.disabled && !params.adminView) {
       redirect(controller:"home", action:"index")
